refactor 패키징 변경 및 변수 네이밍 변경
 i -> index
 의존성의 흐름에 맞게 패키징 변경
