+++ src/main/java/calculator/domain/operator/OperationStarategy.java
@@ -1,4 +1,4 @@
-package calculator.operator;
+package calculator.domain.operator;
 
 @FunctionalInterface
 public interface OperationStarategy {
