+++ src/main/java/calculator/domain/operator/OperatorGroup.java
@@ -1,4 +1,4 @@
-package calculator.operator;
+package calculator.domain.operator;
 
 import java.util.Arrays;
 import java.util.NoSuchElementException;
